Config: Use u32 for config version
This brings CachedValue<T> down to 64 bits in size in many cases
(the notable exception being T = std::string), which is useful
for AtomicCachedValue<T>.

I've added a new way of comparing config versions in order
to make the config system not break on version wraparound.

+++ Source/Core/Common/Config/Config.cpp
@@ -18,7 +18,7 @@ using Layers = std::map<LayerType, std::shared_ptr<Layer>>;
 static Layers s_layers;
 static std::list<ConfigChangedCallback> s_callbacks;
 static u32 s_callback_guards = 0;
-static std::atomic<u64> s_config_version = 0;
+static std::atomic<u32> s_config_version = 0;
 
 static std::shared_mutex s_layers_rw_lock;
 
@@ -83,7 +83,7 @@ void OnConfigChanged()
     callback();
 }
 
-u64 GetConfigVersion()
+u32 GetConfigVersion()
 {
   return s_config_version.load(std::memory_order_relaxed);
 }
